added bin utilities into the lib
